Add open file matching pattern and minor other changes
The OpenFileMatchingPattern is added allowing switching between multiple
files. To achieve this the api needed to be changed of the setting
views, to allow a single view to return multiple action names and
shortcuts to be verified.

Also added the option to search with the previous search querry again
after re-opening the search view